refactor(web): 媒体库表单弹窗从 library-view 拆到 library-form-dialog
首页文件只剩浏览逻辑；LibraryFormDialog 及收藏范围表单辅助函数搬入
components/library-form-dialog.tsx，单库页改从新文件导入。行为不变。
routingOverlapWarnings 改为导出，供后续管理页复用。
